requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
openpyxl==3.1.2
streamlit==1.28.2
PyPDF2==3.0.1
pdfplumber==0.10.3
anthropic==0.34.0
python-docx==0.8.11
python-pptx==0.6.23
# 共有データベース用
psycopg2-binary==2.9.7
toml==0.10.2 